Add comprehensive tests for memory info parsing, settings serialization, color conversion, file path handling, formatting utilities, IP address validation, and rounding functions
- Implement tests for parsing memory total from various formats in `procfs.rs`
- Enhance settings serialization tests in `settings.rs` to cover edge cases
- Add tests for hex to RGB conversion in `color.rs`
- Extend file path handling tests in `file.rs` for different environments and edge cases
- Introduce tests for formatting utilities in `formatter.rs`
- Validate IP address handling in `ip.rs` with link-local checks
- Add extensive rounding tests in `rounding.rs` for various scenarios
