+++ src-tauri/src/_tests/commands/mod.rs
@@ -2,5 +2,3 @@
 pub mod background_image_test;
 #[cfg(test)]
 pub mod settings_test;
-//#[cfg(test)]
-//pub mod hardware_test;
