--- src-tauri/src/_tests/enums/mod.rs
@@ -1,3 +1,0 @@
-pub mod error_test;
-pub mod hardware_test;
-pub mod settings_test;
