--- src-tauri/src/_tests/infrastructure/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(test)]
-pub mod providers;
