--- src-tauri/src/_tests/infrastructure/providers/linux/mod.rs
@@ -1,11 +1,0 @@
-#[cfg(target_os = "linux")]
-#[cfg(test)]
-pub mod dmidecode_test;
-
-#[cfg(target_os = "linux")]
-#[cfg(test)]
-pub mod kernel_test;
-
-#[cfg(target_os = "linux")]
-#[cfg(test)]
-pub mod procfs_test;
