--- src-tauri/src/_tests/infrastructure/providers/mod.rs
@@ -1,3 +1,0 @@
-#[cfg(target_os = "linux")]
-#[cfg(test)]
-pub mod linux;
