+++ src-tauri/src/_tests/mod.rs
@@ -1,14 +1,2 @@
 #[cfg(test)]
 pub mod commands;
-
-#[cfg(test)]
-pub mod enums;
-
-#[cfg(test)]
-pub mod infrastructure;
-
-#[cfg(test)]
-pub mod models;
-
-#[cfg(test)]
-pub mod utils;
