--- src-tauri/src/_tests/models/mod.rs
@@ -1,1 +1,0 @@
-pub mod settings_test;
