--- src-tauri/src/_tests/services/mod.rs
@@ -1,1 +1,0 @@
-pub mod memory_service_test;
