--- src-tauri/src/_tests/utils/mod.rs
@@ -1,11 +1,0 @@
-#[cfg(test)]
-pub mod color_test;
-#[cfg(test)]
-pub mod file_test;
-#[cfg(test)]
-pub mod formatter_test;
-#[cfg(test)]
-#[cfg(target_os = "windows")]
-pub mod ip_test;
-#[cfg(test)]
-pub mod rounding_test;
